refactor(web): move env-layout restore helpers into restore module
dockview-desktop-layout.tsx had grown to 630 lines, 30 over the project's
soft cap. Move `collectPhantomSessionIdsForEnv` and `restoreEnvLayout` to
`dockview-layout-restore.ts` where `tryRestoreLayout` lives; the layout
component now passes `VALID_COMPONENTS` in. Tests for the phantom helper
move alongside, keeping the boundary intact.
