+++ apps/web/components/task/dockview-desktop-layout.test.ts
@@ -1,9 +1,5 @@
 import { describe, it, expect } from "vitest";
-import {
-  CHAT_PANEL_FALLBACK_LABEL,
-  collectPhantomSessionIdsForEnv,
-  resolveChatPanelTitle,
-} from "./dockview-desktop-layout";
+import { CHAT_PANEL_FALLBACK_LABEL, resolveChatPanelTitle } from "./dockview-desktop-layout";
 
 /**
  * Regression: the generic "chat" placeholder dockview panel used to fall back
@@ -34,38 +30,5 @@ describe("resolveChatPanelTitle", () => {
     for (const name of ["Mock", "Claude Code", "GPT-5", "amp"]) {
       expect(resolveChatPanelTitle(name)).toBe(name);
     }
-  });
-});
-
-/**
- * Regression: an env-layout could be restored with a session panel whose id
- * referred to a previously-deleted task's session. The fix strips session
- * panels we KNOW belong to a different env; sessions not yet mapped in the
- * store are preserved (they may still be loading via WS).
- */
-describe("collectPhantomSessionIdsForEnv", () => {
-  it("returns session ids whose mapping is a different env", () => {
-    const state = {
-      environmentIdBySessionId: {
-        "sess-1": "env-A",
-        "sess-2": "env-A",
-        "sess-3": "env-B",
-      },
-    };
-    expect(collectPhantomSessionIdsForEnv(state, "env-A")).toEqual(new Set(["sess-3"]));
-    expect(collectPhantomSessionIdsForEnv(state, "env-B")).toEqual(new Set(["sess-1", "sess-2"]));
-  });
-
-  it("returns every mapped session as phantom when the env has no own sessions yet", () => {
-    const state = { environmentIdBySessionId: { "sess-1": "env-A" } };
-    expect(collectPhantomSessionIdsForEnv(state, "env-new")).toEqual(new Set(["sess-1"]));
-  });
-
-  it("does NOT classify a session as a phantom when its mapping is absent (still loading via WS)", () => {
-    // A session id present in a saved layout but not in environmentIdBySessionId
-    // could be a not-yet-arrived session for this very env. Keep it; reconcile
-    // will clean it up later if it really is stale.
-    const state = { environmentIdBySessionId: {} };
-    expect(collectPhantomSessionIdsForEnv(state, "env-A")).toEqual(new Set());
   });
 });
